% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dola.R
\name{dola}
\alias{dola}
\title{DoLa: Do Lattes para o PPG}
\usage{
dola(
  id_especifico = NULL,
  ano_ini,
  ano_fim,
  nome_instituicao,
  nome_curso,
  nome_area,
  nome_comissao = NULL,
  xlsx_qualis,
  xlsx_qualis_livros = NULL,
  doce_cv,
  dice_cv,
  fl_classificacao
)
}
\arguments{
\item{id_especifico}{description}

\item{ano_ini}{description}

\item{ano_fim}{description}

\item{nome_instituicao}{description}

\item{nome_curso}{description}

\item{nome_area}{description}

\item{nome_comissao}{description}

\item{xlsx_qualis}{description}

\item{xlsx_qualis_livros}{description}

\item{doce_cv}{description}

\item{dice_cv}{description}

\item{fl_classificacao}{description}
}
\description{
esta função extrai dados do curriculo Lattes dos docentes
de um determinado Programa de Pós-graduação
}
